requirement.txt
absl-py==0.15.0
ale-py
astunparse==1.6.3
certifi==2022.9.14
charset-normalizer==2.1.1
clang==5.0
cycler==0.11.0
decorator==5.1.1
dm-control
dm-env==1.5
dm-tree==0.1.7
flatbuffers==1.12
fonttools
gast==0.4.0
glfw==2.5.5
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
gym==0.18.0
gym-notices==0.0.8
h5py==3.1.0
idna
importlib-metadata
importlib-resources
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver
labmaze==1.0.5
lxml==4.9.1
Markdown
MarkupSafe
matplotlib
multipledispatch==0.6.0
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==7.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
pyglet==1.5.0
PyOpenGL==3.1.6
pyparsing==2.4.7
pyrr
requests
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scipy
six==1.15.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.0
tensorflow-estimator
tensorflow-probability==0.14.0
termcolor==1.1.0
tqdm==4.64.1
typing-extensions==3.7.4.3
urllib3
Werkzeug
wrapt==1.12.1
zipp
